Fix Hugo module init and readme
Motivation:
'hugo mod init' fails as-is and requires a workaround (removing the theme dependency).

Modifications:
Commit the go.mod file after successfully initializing the Hugo module. Also add the prerequisites to the readme file.

Result:
The commands mentioned in the readme should work as expected after checking out the repo.